requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
attrs=19.3.0=py_0
backcall=0.1.0=py37_0
blas=1.0=mkl
bleach=3.1.0=py37_0
ca-certificates=2019.11.28=hecc5488_0
certifi=2019.11.28=py37hc8dfbb8_1
cffi=1.14.0=py37h7a1dbc1_0
chardet=3.0.4=pypi_0
cloudpickle=1.3.0=py_0
colorama=0.4.3=py_0
correlation-cuda=0.0.0=pypi_0
cudatoolkit=10.1.243=h74a9793_0
cycler=0.10.0=py37_0
cytoolz=0.10.1=py37hfa6e2cd_0
dask-core=2.13.0=py_0
decorator=4.4.1=py_0
defusedxml=0.6.0=py_0
efficientnet-pytorch=0.6.3=pypi_0
entrypoints=0.3=py37_0
facenet-pytorch=2.2.7=pypi_0
freetype=2.9.1=ha9979f8_1
hdf5=1.8.20=hac2f561_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
idna=2.9=pypi_0
imageio=2.8.0=py_0
imgaug=0.4.0=pypi_0
importlib_metadata=1.5.0=py37_0
intel-openmp=2020.0=166
ipykernel=5.1.4=py37h39e3cac_0
ipython=7.12.0=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.1=py_0
jedi=0.16.0=py37_0
jinja2=2.11.1=py_0
joblib=0.14.1=py_0
jpeg=9c=hfa6e2cd_1001
jsonschema=3.2.0=py37_0
jupyter_client=5.3.4=py37_0
jupyter_core=4.6.1=py37_0
kiwisolver=1.1.0=py37ha925a31_0
libblas=3.8.0=15_mkl
libcblas=3.8.0=15_mkl
libclang=9.0.1=default_hf44288c_0
liblapack=3.8.0=15_mkl
liblapacke=3.8.0=15_mkl
libopencv=3.4.2=h20b85fd_0
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.16=h9d3ae62_0
libtiff=4.1.0=h56a325e_0
libwebp=1.0.2=hfa6e2cd_5
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=1.1.1=py37he774522_0
matplotlib=3.1.3=py37_0
matplotlib-base=3.1.3=py37h64f37c6_0
mistune=0.8.4=py37he774522_0
mkl=2020.0=166
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.15=py37h14836fe_0
mkl_random=1.1.0=py37h675688f_0
msys2-conda-epoch=20160418=1
nbconvert=5.6.1=py37_0
nbformat=5.0.4=py_0
networkx=2.4=py_1
ninja=1.9.0=py37h74a9793_0
notebook=6.0.1=py37_0
numpy=1.18.1=py37h93ca92e_0
numpy-base=1.18.1=py37hc3f5095_1
olefile=0.46=py37_0
opencv=3.4.2=py37h40b0b35_0
opencv-python=4.2.0.34=pypi_0
openssl=1.1.1e=hfa6e2cd_0
pandas=1.0.1=py37h47e9c7a_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py37_1
parso=0.6.1=py_0
pickleshare=0.7.5=py37_0
pillow=7.0.0=py37hcc1f983_0
pip=20.0.2=py37_1
prometheus_client=0.7.1=py_0
prompt_toolkit=3.0.3=py_0
py-opencv=3.4.2=py37hc319ecb_0
pycparser=2.19=py37_0
pygments=2.5.2=py_0
pyparsing=2.4.6=py_0
pypng=0.0.20=pypi_0
pyqt=5.9.2=py37h6538335_2
pyrsistent=0.15.7=py37he774522_0
python=3.7.6=h60c2a47_2
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.4.0=py3.7_cuda101_cudnn7_0
pytorch-gradcam=0.2.1=pypi_0
pytube3=9.6.4=pypi_0
pytz=2019.3=py_0
pywavelets=1.1.1=py37hc8d92b1_0
pywin32=227=py37he774522_1
pywinpty=0.5.7=py37_0
pyzmq=18.1.1=py37ha925a31_0
qt=5.9.7=vc14h73c81de_0
requests=2.23.0=pypi_0
scikit-image=0.16.2=py37he350917_0
scikit-learn=0.22.1=py37h6288b17_0
scipy=1.4.1=py37h9439919_0
seaborn=0.10.0=pypi_0
send2trash=1.5.0=py37_0
setuptools=45.2.0=py37_0
shapely=1.7.0=pypi_0
sip=4.19.8=py37h6538335_0
six=1.14.0=py37_0
sqlite=3.31.1=he774522_0
terminado=0.8.3=py37_0
testpath=0.4.4=py_0
tk=8.6.8=hfa6e2cd_0
toolz=0.10.0=py_0
torchvision=0.5.0=py37_cu101
tornado=6.0.3=py37he774522_3
tqdm=4.42.1=py_0
traitlets=4.3.3=py37_0
typing-extensions=3.7.4.2=pypi_0
urllib3=1.25.8=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_1
wcwidth=0.1.8=py_0
webencodings=0.5.1=py37_1
wheel=0.34.2=py37_0
widgetsnbextension=3.5.1=py37_0
wincertstore=0.2=py37_0
winpty=0.4.3=4
xz=5.2.4=h2fa13f4_4
zeromq=4.3.1=h33f27b4_3
zipp=2.2.0=py_0
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
